+++ dql/util.py
@@ -1,5 +1,4 @@
 """ Utility methods """
-import six
 from decimal import Decimal
 from dynamo3 import Binary
 
@@ -33,32 +32,3 @@ def resolve(val):
         return dict_val
     else:
         raise SyntaxError("Unable to resolve value '%s'" % val)
-
-
-def plural(value, append='s'):
-    """ Helper function for pluralizing text """
-    return '' if value == 1 else append
-
-
-def pretty_format(statement, result):
-    """ Format the return value of a query for humans """
-    if result is None:
-        return 'Success'
-    if statement.action == 'UPDATE':
-        if isinstance(result, six.integer_types):
-            return "Updated %d item%s" % (result, plural(result))
-    elif statement.action == 'DELETE':
-        return "Deleted %d item%s" % (result, plural(result))
-    elif statement.action == 'CREATE':
-        if result:
-            return "Created table %r" % statement.table
-        else:
-            return "Table %r already exists" % statement.table
-    elif statement.action == 'INSERT':
-        return "Inserted %d item%s" % (result, plural(result))
-    elif statement.action == 'DROP':
-        if result:
-            return "Dropped table %r" % statement.table
-        else:
-            return "Table %r does not exist" % statement.table
-    return result